modules/requirements.txt
PyNaCl>=1.5.0
requests>=2.32.5
